fix: themes now work thankfully
